feat(PRL-3063) - Replacing current dashboard with new account dashboard
