+++ src/main/routes.ts
@@ -19,7 +19,6 @@ import ApplicantConfirmContactDetailsPostController from './steps/applicant/conf
 import { SupportYouNeedDuringYourCaseController } from './steps/applicant/support-you-need-during-case/SupportYouNeedDuringCaseController';
 import { ApplicationDownloadController } from './steps/c100-rebuild/confirmation-page/ApplicationDownloadController';
 import { ViewAllDocumentsPostController } from './steps/common/controller/ViewAllDocumentsPostController';
-import DashboardGetController from './steps/common/dashboard/DashboardGetController';
 import { KeepDetailsPrivateGetController } from './steps/common/keep-details-private/KeepDetailsPrivateGetController';
 import { KeepDetailsPrivatePostController } from './steps/common/keep-details-private/KeepDetailsPrivatePostController';
 import CaseDetailsGetController from './steps/common/task-list/controllers/CaseDetailsGetController';
@@ -28,6 +27,7 @@ import { ContactUsGetController } from './steps/contact-us/get';
 import { CookiesGetController } from './steps/cookies/get';
 import { ErrorController } from './steps/error/error.controller';
 import { PrivacyPolicyGetController } from './steps/privacy-policy/get';
+import DashboardGetController from './steps/prl-cases/dashboard/DashboardGetController';
 import { RespondentConfirmContactDetailsGetController } from './steps/respondent/confirm-contact-details/checkanswers/controller/RespondentConfirmContactDetailsGetController';
 import RespondentConfirmContactDetailsPostController from './steps/respondent/confirm-contact-details/checkanswers/controller/RespondentConfirmContactDetailsPostController';
 import { ConsentGetController } from './steps/respondent/consent-to-application/ConsentGetController';
@@ -106,7 +106,6 @@ import {
   CA_DA_SUPPORT_YOU_NEED_DURING_CASE_SAVE,
   C7_SUPPORT_YOU_NEED_DURING_CASE_SAVE,
   RESPONDENT_CHECK_ANSWERS_NO,
-  DASHBOARD_URL_NEW,
   FETCH_CASE_DETAILS,
   PARTY_TASKLIST,
   //C100_DOCUMENT_SUBMISSION,
@@ -119,7 +118,7 @@ export class Routes {
 
     app.get(CSRF_TOKEN_ERROR_URL, errorHandler(errorController.CSRFTokenError));
     app.get(HOME_URL, (req, res) => res.redirect(DASHBOARD_URL));
-    app.get(DASHBOARD_URL_NEW, errorHandler(new DashboardGetController().get));
+    app.get(DASHBOARD_URL, errorHandler(new DashboardGetController().get));
     app.get(FETCH_CASE_DETAILS, errorHandler(new CaseDetailsGetController().get));
     app.get(PARTY_TASKLIST, errorHandler(new TaskListGetController().get));
     app.get(COOKIES_PAGE, errorHandler(new CookiesGetController().get));
