--- src/main/steps/common/dashboard/template.njk
@@ -1,55 +1,0 @@
-{% from "govuk/components/tabs/macro.njk" import govukTabs %}
-{% from "govuk/components/table/macro.njk" import govukTable %}
-{% extends "../page.njk" %}
-
-{% block backLink %}{% endblock %}
-
-{% block content %}
-  {% block page_content %}
-    <h1 class="govuk-heading-l govuk-!-margin-bottom-5">
-      <span class="govuk-caption-l">{{additionalData.user.fullname}}</span>
-      {{title}}
-    </h1>
-
-    <h3 class="govuk-heading-s">{{sectionTitle}}</h3>
-
-    <ul class="govuk-list">
-      <li>
-        <a class="govuk-link" href="/citizen-home">{{activateAccessCodeLinkText}}</a>
-      </li>
-      <li>
-        <a class="govuk-link" href="/task-list/applicant">{{newChildArrangementsLinkText}}</a>
-      </li>
-      <li>
-        <a class="govuk-link" href="https://www.gov.uk/injunction-domestic-violence">{{newFamilyInjunctionLinkText}}</a>
-      </li>
-    </ul>
-
-    {% set tabItems = [] %}
-    {% for caseState, tab in tabs %}
-      {% set panelContent %}
-      <h2 class="govuk-heading-m">{{tab.heading}}</h2>
-      {% if tab.body %}
-        <p class="govuk-body">{{tab.body | safe}}</p>
-      {% endif %}
-      {{ govukTable({
-          head: tab.head,
-          rows: tab.rows
-        }) }}
-      {% endset %}
-      {% set tabItems = (tabItems.push(
-        {
-          label: tab.label,
-          id: tab.id,
-          panel: {
-            html: panelContent
-          }
-        }
-      ), tabItems) %}
-    {% endfor %}
-    {{ govukTabs({
-      classes: 'govuk-!-margin-top-8',
-      items: tabItems
-    }) }}
-  {% endblock %}
-{% endblock %}
