--- src/main/steps/prl-cases/dashboard-v1/content.ts
@@ -1,23 +1,0 @@
-import { TranslationFn } from '../../../app/controller/GetController';
-
-const en = () => ({
-  title: 'Welcome to Citizen dashboard',
-  paymentError: '',
-});
-
-const cy = () => ({
-  title: 'Welcome to Citizen dashboard(welsh)',
-  paymentError: '',
-});
-
-const languages = {
-  en,
-  cy,
-};
-
-export const generateContent: TranslationFn = content => {
-  const translations = languages[content.language]();
-  return {
-    ...translations,
-  };
-};
