+++ src/main/steps/prl-cases/dashboard/DashboardGetController.ts
@@ -1,34 +1,32 @@
 import autobind from 'autobind-decorator';
 import { Response } from 'express';
 
-import { getCaseDetails } from '../../../../main/app/auth/user/oidc';
-import { AppRequest } from '../../../../main/app/controller/AppRequest';
-import { GetController, TranslationFn } from '../../../../main/app/controller/GetController';
-import { Language, generatePageContent } from '../../../../main/steps/common/common.content';
+import { getCaseDetails } from '../../../app/auth/user/oidc';
+import { AppRequest } from '../../../app/controller/AppRequest';
+import { GetController } from '../../../app/controller/GetController';
+
+import { generateContent } from './content';
 
 @autobind
 export default class DashboardGetController extends GetController {
-  constructor(protected readonly view: string, protected readonly content: TranslationFn) {
-    super(view, content);
+  constructor() {
+    super(`${__dirname}/template`, generateContent);
   }
 
   public async get(req: AppRequest, res: Response): Promise<void> {
-    req.session.userCaseList = await getCaseDetails(req);
-
-    const language = super.getPreferredLanguage(req) as Language;
-    const content = generatePageContent({
-      language,
-      pageContent: this.content,
-      userCase: req.session?.userCase,
-      userCaseList: req.session?.userCaseList,
-      additionalData: {
-        req,
-      },
-    });
-    res.render(this.view, {
-      ...content,
-      htmlLang: language,
-      userIdamId: req.session?.user?.id,
-    });
+    try {
+      req.session.userCaseList = await getCaseDetails(req);
+      clean(req.session);
+      req.session.save(() => {
+        super.get(req, res);
+      });
+    } catch (e) {
+      super.get(req, res);
+    }
   }
+}
+
+function clean(session) {
+  delete session.userCase;
+  return session;
 }
