--- src/main/steps/prl-cases/dashboard/customSummaryList/macro.njk
@@ -1,3 +1,0 @@
-{% macro govukSummaryList(params) %}
-  {%- include "./template.njk" -%}
-{% endmacro %}
