+++ src/main/steps/prl-cases/dashboard/template.njk
@@ -1,22 +1,55 @@
+{% from "govuk/components/tabs/macro.njk" import govukTabs %}
+{% from "govuk/components/table/macro.njk" import govukTable %}
 {% extends "../../common/dashboard-page.njk" %}
 
-{% block page_content %}
-  {% if section %}
-    <span class="govuk-heading-l">{{ section }}</span>
-  
-  {% endif %}
-  {% if title %}
-    <h1>{{ title }}</h1>
-  {% endif %}
-  {% if line1 %}
-    <div class="govuk-hint">{{ line1 }}</div>
-  {% endif %}
-  {% if line2 %}
-    <p class="govuk-body">{{ line2 }}</p>
-  {% endif %}
+{% block backLink %}{% endblock %}
 
-{% endblock %}
+{% block content %}
+  {% block page_content %}
+    <h1 class="govuk-heading-l govuk-!-margin-bottom-5">
+      <span class="govuk-caption-l">{{additionalData.user.fullname}}</span>
+      {{title}}
+    </h1>
 
-{% block form %}
-  {% include "./form.njk"%}
+    <h3 class="govuk-heading-s">{{sectionTitle}}</h3>
+
+    <ul class="govuk-list">
+      <li>
+        <a class="govuk-link" href="/citizen-home">{{activateAccessCodeLinkText}}</a>
+      </li>
+      <li>
+        <a class="govuk-link" href="/task-list/applicant">{{newChildArrangementsLinkText}}</a>
+      </li>
+      <li>
+        <a class="govuk-link" href="https://www.gov.uk/injunction-domestic-violence">{{newFamilyInjunctionLinkText}}</a>
+      </li>
+    </ul>
+
+    {% set tabItems = [] %}
+    {% for caseState, tab in tabs %}
+      {% set panelContent %}
+      <h2 class="govuk-heading-m">{{tab.heading}}</h2>
+      {% if tab.body %}
+        <p class="govuk-body">{{tab.body | safe}}</p>
+      {% endif %}
+      {{ govukTable({
+          head: tab.head,
+          rows: tab.rows
+        }) }}
+      {% endset %}
+      {% set tabItems = (tabItems.push(
+        {
+          label: tab.label,
+          id: tab.id,
+          panel: {
+            html: panelContent
+          }
+        }
+      ), tabItems) %}
+    {% endfor %}
+    {{ govukTabs({
+      classes: 'govuk-!-margin-top-8',
+      items: tabItems
+    }) }}
+  {% endblock %}
 {% endblock %}
